requirements.txt
boto3
opensearch-py
python-dotenv
telethon